add a flow between two nodes in the x-direction

add a flow between two nodes in the x-direction

flowx(
  from = NULL,
  to = NULL,
  length = NULL,
  label = NULL,
  pos = NULL,
  label_pos = NULL,
  label_gap = NULL,
  label_x = NULL,
  label_y = NULL,
  label_col = "black",
  label_font = 3,
  label_cex = 0.8,
  arr_type = "triangle",
  arr_col = "black",
  arr_length = 0.15,
  arr_width = NULL,
  arr_lty = 1,
  name_from = deparse(substitute(from)),
  name_to = deparse(substitute(to)),
  ...
)

Arguments

from

start node

to

end node

length

single numeric specifying length, positive will result in a flow from left to right, negative from right to left. Optional parameter replacing either to or from

label

label to draw on flow

pos

a decimal between 0 and 1 giving the position of the flow relative to the nodes. Where 0 = bottom and 1 = top of the interval over which the nodes overlap.

label_pos

a decimal between 0 and 1 giving the x-position along the flow to draw the label, where 0 = left and 1 = right

label_gap

y-distance from the flow at which to draw the label, default = 0.05

label_x

x co-ordinate of label position, overrides use of label_pos

label_y

y co-ordinate of label position, overrides use of label_gap

label_col

colour of label, defaults to black

label_font

font of label, defaults to 3

label_cex

label cex, defaults to 0.8

arr_type

type of arrow head to draw default = triangle

arr_col

colour of arrow default = black

arr_length

length of arrow head, defaults to 0.2

arr_width

width of arrow head, defaults to 0.1

arr_lty

lty of arrow, defaults to 1

name_from

internal argument used for informative error messages

name_to

internal argument used for informative error messages

...

additional formatting arguments to arrow()

Value

returns the start and end points of the flow