add fork between three nodes, in the x-direction

add fork between three nodes, in the x-direction

forkx(
  from = NULL,
  to0,
  to1,
  length = NULL,
  pos = NULL,
  label_from = NULL,
  label_to0 = NULL,
  label_to1 = NULL,
  pos_from = NULL,
  pos_to0 = NULL,
  pos_to1 = pos_to0,
  label_from_pos = NULL,
  label_to0_pos = label_from_pos,
  label_to1_pos = label_to0_pos,
  label_from_gap = NULL,
  label_to0_gap = label_from_gap,
  label_to1_gap = label_to0_gap,
  label_from_x = NULL,
  label_to0_x = NULL,
  label_to1_x = label_to0_x,
  label_from_y = NULL,
  label_to0_y = NULL,
  label_to1_y = NULL,
  arr_width = NULL,
  name_from = deparse(substitute(from)),
  name_to0 = deparse(substitute(to0)),
  name_to1 = deparse(substitute(to1)),
  ...
)

Arguments

from
start node

to0
the bottom end node

to1
the top end node

length
single numeric specifying length, positive will result in a fork from left to right, negative from right to left. Optional parameter replacing from

pos
a decimal between 0 and 1 giving the position between from, and the closest of to0 and to1 at which to split the fork.

label_from
label to draw on the flow out of from

label_to0
label to draw on the flow into to0

label_to1
label to draw on the flow into to1

pos_from
a decimal between 0 and 1 giving the position to start the flow out of from, where 0 = bottom and 1 = top

pos_to0
a decimal between 0 and 1 giving the position to attach the flow to node to0, where 0 = bottom and 1 = top

pos_to1
a decimal between 0 and 1 giving the position to attach the flow to node to1, where 0 = bottom and 1 = top

label_from_pos
a decimal between 0 and 1 giving the position along the flow out of from to draw label_from, where 0 = start and 1 = end

label_to0_pos
a decimal between 0 and 1 giving the position along the flow into to0 to draw label_to0, where 0 = start and 1 = end

label_to1_pos
a decimal between 0 and 1 giving the position along the flow into to1 to draw label_to1, where 0 = start and 1 = end

label_from_gap
distance from the flow out of from at which to draw label_from default = 0.05

label_to0_gap
distance from the flow into to0 at which to draw label_to0 default = 0.05

label_to1_gap
distance from the flow into to1 at which to draw label_to1 default = 0.05

label_from_x
x co-ordinate of label_from position, overrides use of label_from_pos and / or label_from_gap

label_to0_x
x co-ordinate of label_to0 position, overrides use of label_to0_pos and / or label_to0_gap

label_to1_x
x co-ordinate of label_to1 position, overrides use of label_to1_pos and / or label_to1_gap

label_from_y
y co-ordinate of label_from position, overrides use of label_from_pos and / or label_from_gap

label_to0_y
y co-ordinate of label_to0 position, overrides use of label_to0_pos and / or label_to0_gap

label_to1_y
y co-ordinate of label_to1 position, overrides use of label_to1_pos and / or label_to1_gap

arr_width
width of arrow, defaults to same as flow()

name_from
internal argument used for informative error messages

name_to0
internal argument used for informative error messages

name_to1
internal argument used for informative error messages

...
additional formatting arguments to flow()

Value

returns the start and end points of the flow