group a set of nodes and flows

group a set of nodes and flows

group(
  f,
  args = list(),
  oma = rep(0.1, 4),
  group_col = NULL,
  border_col = "black",
  label = "",
  label_pos_x = 0.02,
  label_pos_y = 0.9,
  label_font = 1,
  label_col = "black",
  label_cex = 1,
  label_x = NULL,
  label_y = NULL,
  ...
)

Arguments

f
flodia plot function

args
list of named arguments to f

oma
required border around nodes, vector of length 4 specifying: bottom, left, top, right. defaults to 0.1 on all sides

group_col
colour of group box

border_col
colour of group border

label
label to draw on group

label_pos_x
a decimal between 0 and 1 giving the position on the group to draw label, where 0 = left and 1 = right

label_pos_y
a decimal between 0 and 1 giving the position on the group to draw label, where 0 = bottom and 1 = top

label_font
font of label, defaults to 1

label_col
colour of label, defaults to black

label_cex
label cex, defaults to 1

label_x
x co-ordinate of label position, overrides label_pos_x

label_y
y co-ordinate of label position, overrides label_pos_y

...
further arguments to rect() for formatting group

Value

coordinates of group